Move Dto annotation to the core module
